feat: :tada: updates CLI version to v0.1.1
Increments the CLI version to v0.1.1.

This commit also includes updates to the SVG animation generation process by using `make svg-animation` command.
It adds Go version setup in the workflow for generating SVG animations, and updates the go version in `go.mod` to 1.25.
The scripts are also adjusted to execute the binary locally.